Extract class Ui
